+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [BREAKIN CHANGE] Migrate `SelectWithState` from `react-select@1` to `react-select@3`
+
+  - Our composant follow closer the new `react-select` API
+  - We have removed this following props : `onInputChange`, `deleteRemoves`, `multi`, `inputProps` and `labelText`
+  - Use `Field.Select` from `@kisskissbankbank/kitten/src/next`. `SelectWithState` import has been deprecated.
+  - Styles strategy has changed, check our storybook https://kisskissbankbank.github.io/kitten for more details
+
 ## [2.93.0] - 2020-09-30
 
 Features: