+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `micro` size for `<Button icon>`.
+
 ## [2.94.0] - 2020-10-13
 
 Feature: